requirements.txt
psycopg2
fastapi
uvicorn
pathlib
pyyaml